[utils] Minor refactoring: move `processedConfig.js` functions to utils.js
